Split simplified user interface components into new file driconf_simpleui.py.
Implemented a very simple initial version of the simplified user
interface that can only change the default app config of the first
screen in the user configuration file. The configuration file is saved
automatically on every change.

Removed unnecessary callbacks to app.modified in SectionPage.commit that
lead to endless recursion in the simplified UI.

Always start the simplified user interface for now for testing.