Added options -e to start in expert mode.
Added simple and removable parameters to OptionLine constructor.
Removable turns the reset-button into a remove-button. Added tooltip to
reset/remove button.

Added application configuration inside an Expander to the simple UI.
Applications are selected using a ComboBox. Added an AppPage
(ScrolledWindow) that represents application option settings. Unknown
options are not represented yet. OptionLines are removable. In the
bottom there is a ComboBox for adding new options (not functional yet).